M: ping_message added. some rocket cleanup int32_t synced on damage_taken.

+++ src/network/damage_taken_message.h
@@ -13,13 +13,13 @@ namespace network {
     public:
         DEFINE_MESSAGE_BASE(DamageTakenMessage)
         int32_t mActorGUID;
-        double mX;
-        double mY;
+        int32_t mX;
+        int32_t mY;
         int32_t mDamage;
         DamageTakenMessage()
             : mActorGUID(0)
-            , mX(0.0)
-            , mY(0.0)
+            , mX(0)
+            , mY(0)
             , mDamage(0)
         {
         }
